Fixes to CR comments
